+++ data-plane-aas/src/main/java/de/fraunhofer/iosb/aas/AasDataProcessorFactory.java
@@ -16,7 +16,11 @@
 package de.fraunhofer.iosb.aas;
 
 import de.fraunhofer.iosb.ssl.SelfSignedCertificateRetriever;
-import org.eclipse.edc.spi.EdcException;
+import dev.failsafe.RetryPolicy;
+import okhttp3.OkHttpClient;
+import okhttp3.Response;
+import org.eclipse.edc.http.client.EdcHttpClientImpl;
+import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.spi.result.Result;
 
 import java.net.MalformedURLException;
@@ -25,15 +29,25 @@ import java.security.KeyStoreException;
 import java.security.cert.Certificate;
 
 import static de.fraunhofer.iosb.aas.http.HttpClientProvider.clientFor;
-import static java.lang.String.format;
 
 public abstract class AasDataProcessorFactory {
 
     private static final String HTTPS = "HTTPS";
     protected final SelfSignedCertificateRetriever retriever;
+    // EDC provided fields. Use for non-self-signed certificates
+    private final OkHttpClient edcOkHttpClient;
+    private final RetryPolicy<Response> edcRetryPolicy;
+    private final Monitor monitor;
 
-    public AasDataProcessorFactory(SelfSignedCertificateRetriever retriever) {
+    public AasDataProcessorFactory(SelfSignedCertificateRetriever retriever,
+                                   OkHttpClient edcOkHttpClient,
+                                   RetryPolicy<Response> edcRetryPolicy,
+                                   Monitor monitor) {
         this.retriever = retriever;
+
+        this.monitor = monitor;
+        this.edcOkHttpClient = edcOkHttpClient;
+        this.edcRetryPolicy = edcRetryPolicy;
     }
 
     /**
@@ -44,20 +58,24 @@ public abstract class AasDataProcessorFactory {
      * @param urlString URL of AAS service.
      * @return AAS Processor allowing communication with AAS service using AAS data addresses
      */
-    public AasDataProcessor processorFor(String urlString) {
+    public Result<AasDataProcessor> processorFor(String urlString) {
         URL aasUrl;
         try {
             aasUrl = new URL(urlString);
         } catch (MalformedURLException malformedUrlException) {
-            throw new EdcException("Malformed URL for AAS manipulator", malformedUrlException);
+            return Result.failure(malformedUrlException.getMessage());
         }
 
         var certResult = getCertificates(aasUrl);
 
         try {
-            return new AasDataProcessor(clientFor(certResult.getContent()));
-        } catch (KeyStoreException e) {
-            throw new EdcException(e);
+            return Result.success(new AasDataProcessor(
+                    new EdcHttpClientImpl(
+                            certResult.getContent() == null ? edcOkHttpClient : clientFor(certResult.getContent()),
+                            edcRetryPolicy,
+                            monitor)));
+        } catch (KeyStoreException keyStoreException) {
+            return Result.failure(keyStoreException.getMessage());
         }
     }
 
@@ -72,8 +90,8 @@ public abstract class AasDataProcessorFactory {
         var certsResult = retriever.getSelfSignedCertificate(url);
 
         if (certsResult.failed() && !certsResult.getFailureMessages().contains("trusted")) {
-            throw new EdcException(format("Certificates were neither trusted nor self-signed: %s",
-                    certsResult.getFailureMessages()));
+            return Result.failure("Certificates were neither trusted nor self-signed: %s"
+                    .formatted(certsResult.getFailureMessages()));
         }
 
         return certsResult;