+++ data-plane-aas/src/main/java/de/fraunhofer/iosb/dataplane/aas/pipeline/AasDataSink.java
@@ -15,10 +15,11 @@
  */
 package de.fraunhofer.iosb.dataplane.aas.pipeline;
 
-import de.fraunhofer.iosb.aas.AasDataProcessor;
+import de.fraunhofer.iosb.aas.AasDataProcessorFactory;
 import de.fraunhofer.iosb.dataplane.aas.spi.AasDataAddress;
 import org.eclipse.edc.connector.dataplane.spi.pipeline.DataSink;
 import org.eclipse.edc.connector.dataplane.spi.pipeline.DataSource;
+import org.eclipse.edc.connector.dataplane.spi.pipeline.StreamFailure;
 import org.eclipse.edc.connector.dataplane.spi.pipeline.StreamResult;
 import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.spi.result.AbstractResult;
@@ -31,19 +32,17 @@ import static java.util.concurrent.CompletableFuture.completedFuture;
 import static org.eclipse.edc.connector.dataplane.spi.pipeline.StreamResult.failure;
 
 /**
- * Retrieve the parts from dataSource and send them given
- * the HttpRequestParams and EdcExtension custom http client.
- * This works similarly to HttpDataSink but uses a custom
- * HttpClient accepting certain self-signed certificates.
- * TODO Since foreign AAS services should have valid certificates
- * TODO this sink will probably not be needed in the future
- * TODO (Unless we want to send data to completely untrusted servers)
+ * Retrieves the parts from dataSource and send them given the aas data processor.
  */
 public class AasDataSink implements DataSink {
 
-    private AasDataProcessor aasDataProcessor;
+    private AasDataProcessorFactory aasDataProcessorFactory;
     private AasDataAddress aasDataAddress;
     private Monitor monitor;
+
+    private AasDataSink() {
+
+    }
 
     @Override
     public CompletableFuture<StreamResult<Object>> transfer(DataSource dataSource) {
@@ -60,14 +59,20 @@ public class AasDataSink implements DataSink {
     }
 
     private StreamResult<Object> transferPart(DataSource.Part part) {
+        var aasDataProcessor = aasDataProcessorFactory.processorFor(aasDataAddress.getBaseUrl());
 
-        monitor.debug(() -> "Executing HTTP request to AAS service: " + aasDataAddress.getBaseUrl());
+        if (aasDataProcessor.failed()) {
+            monitor.severe("Error writing HTTP data %s to endpoint %s:\n%s".formatted(part.name(), aasDataAddress.getBaseUrl(),
+                    aasDataProcessor.getFailureMessages()));
 
-        try (var response = aasDataProcessor.send(aasDataAddress, part)) {
+            return StreamResult.failure(new StreamFailure(aasDataProcessor.getFailureMessages(), StreamFailure.Reason.GENERAL_ERROR));
+        }
+
+        try (var response = aasDataProcessor.getContent().send(aasDataAddress, part)) {
             return StreamResult.success("DataTransfer completed. Response from consumer: " + response.body());
         } catch (IOException e) {
             var errorMessage = "IOException while data transferring to AAS: " + e.getMessage();
-            monitor.severe(() -> errorMessage);
+            monitor.severe(() -> errorMessage, e);
             return StreamResult.error(errorMessage);
         }
 
@@ -89,8 +94,8 @@ public class AasDataSink implements DataSink {
             return this;
         }
 
-        public Builder aasManipulator(AasDataProcessor aasDataProcessor) {
-            dataSink.aasDataProcessor = aasDataProcessor;
+        public Builder aasManipulator(AasDataProcessorFactory aasDataProcessor) {
+            dataSink.aasDataProcessorFactory = aasDataProcessor;
             return this;
         }
 
@@ -100,7 +105,7 @@ public class AasDataSink implements DataSink {
         }
 
         public AasDataSink build() {
-            Objects.requireNonNull(dataSink.aasDataProcessor, "aasManipulator");
+            Objects.requireNonNull(dataSink.aasDataProcessorFactory, "aasManipulator");
             Objects.requireNonNull(dataSink.aasDataAddress, "aasDataAddress");
             return dataSink;
         }