+++ changelog.md
@@ -11,11 +11,13 @@ Compatibility: **Eclipse Dataspace Connector v0.8.0**
 * Accepting self-signed certificates now
   optional (`edc.dataplane.aas.acceptOwnSelfSignedCertificates`, `edc.dataplane.aas.acceptAllSelfSignedCertificates`)
 * SubmodelElementList is now a model element inside the SelfDescription (AASv3 change)
+* Synchronizer rate can now be changed at runtime via configuration updates (time unit still in seconds)
 
 * New extension: data-plane-aas
     * Provides custom AAS data source and data sink
     * Communicate with AAS over AASDataAddress (see above, reference chains)
     * Allow communication with AAS with self-signed TLS certificates (configurable)
+    * Use default EdcHttpClient if not self-signed
 
 **Bugfixes**